+++ NEWS.md
@@ -96,6 +96,7 @@
 * Zoom Left/Right Column commands for keyboard users (#5874)
 * Increase maximum plot size for large, high-DPI displays (#4968; thanks to Jan Gleixner)
 * RStudio Server now uses 2048 bit RSA keys, for secure communication of encrypted credentials between server / session and client
+* Add support for running Shiny applications as background jobs (#5190)
 
 ### Bugfixes